Structure: GetNetCharge accumulates into double, not int
itsCharge is a double; summing it into an int truncated any fractional
net charge. No behaviour change for the integer charges used today.

+++ src/Structure/Imp/Structure.C
@@ -22,7 +22,7 @@ int Structure::GetNuclearCharge() const
 }
 double Structure::GetNetCharge() const
 {
-    int chg=0;
+    double chg=0;
     for(auto b:*this) chg+=b->itsCharge;
     return chg;
 }
